refactor(nbd): move platform-specific code to different files
